+++ Runtime/Sensors/CollisionDetectionStrategy.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace Codetox.Sensors
+{
+    [Flags]
+    public enum CollisionDetectionStrategy
+    {
+        Colliders, TriggerColliders
+    }
+}