+++ Runtime/UI/Billboard.cs
@@ -1,0 +1,12 @@
+﻿using UnityEngine;
+
+namespace Codetox.UI
+{
+    public class Billboard : MonoBehaviour
+    {
+        private void Update()
+        {
+            transform.forward = Camera.main.transform.forward;
+        }
+    }
+}